Bundle vee-validate in library.
