+++ src/components/index.js
@@ -4,3 +4,5 @@ export { default as BSelect } from './BSelect'
 export { default as BContainer } from './BContainer'
 export { default as BHeader } from './BHeader'
 export { default as BToggle } from './BToggle'
+export { default as BValidator } from './BValidator'
+export { default as BInput } from './BInput'
