+++ src/index.js
@@ -1,6 +1,7 @@
 import * as components from './components'
 import { peerDependencies, version } from '../package.json'
 import Semver from 'semver'
+import VeeValidate from 'vee-validate'
 
 const Installer = {
   install (Vue, options = {}) {
@@ -15,6 +16,32 @@ const Installer = {
     for (const key in components) {
       Vue.use(components[key])
     }
+
+    const validationConfig = {
+      errorBagName: 'errors',
+      fieldsBagName: 'fields',
+      delay: 0,
+      locale: 'en',
+      dictionary: null,
+      strict: false,
+      classes: false,
+      classNames: {
+        touched: 'touched', // the control has been blurred
+        untouched: 'untouched', // the control hasn't been blurred
+        valid: 'valid', // model is valid
+        invalid: 'invalid', // model is invalid
+        pristine: 'pristine', // control has not been interacted with
+        dirty: 'dirty' // control has been interacted with
+      },
+      events: 'blur',
+      inject: true,
+      validity: false,
+      aria: true,
+      i18n: null, // the vue-i18n plugin instance
+      i18nRootKey: 'validations' // the nested key under which the validation messsages will be located
+    }
+
+    Vue.use(VeeValidate, validationConfig)
   }
 }
